+++ ops/matmul/index.ts
@@ -1,0 +1,1 @@
+export { matmul, type MatmulArgs } from "./reference.js";